requirements.txt
fastapi
transformers
Pillow
python-multipart
torch
numpy
tensorflow